requirements.txt
tree-sitter
tree-sitter-java
networkx
transformers
torch
sentence-transformers
faiss-cpu
requests